+++ electronics_lib/__init__.py
@@ -106,3 +106,6 @@ from .Bldc_Drv8313 import Drv8313
 
 from .Imu_Lsm6ds3trc import Imu_Lsm6ds3trc
 from .Mag_Qmc5883l import Mag_Qmc5883l
+
+from .LedMatrix import CharlieplexedLedMatrix
+from .SwitchMatrix import SwitchMatrix
